experiments: move testbed/ under datakit/
